Remove excess whitespace from Board.py class
OCD found this on vim with my new plugin. Trying to slowly make the
transition to Vim, but it's a slow and agonizing process so far. Though,
I'm really coming to like it finally that I know a few more important
keybinds like dd, w, e, b, A, u/R, and o/O.