Create additional classes and seperate test file
The testRun() function was seperated to its own file for unit
testing so it will be easier to just run what tests I need.

Create Game.py and Engine.py to house functions for initiating and
playing a game and selecting the best move for a given board state,
respectively.