Remove CLI class and add Group class
In addition to leaving a placeholder for the Group class, I changed
Engine.makeMove to Engine.makeEngineMove to be more clear
about which method makes a move in the Board or Engine classes.
Also made a format change where the -> {data type} at the end of
functions is removed for functions like __init__ or __str__ which are
clear in what data type they do or don't return.

 I also made all variables, methods, etc. convert to snake_case as
 opposed to lowerCamelCase. I do like it better, but since this is an
 open-source project I will try to stand by the PEP8 guidelines.