Overhaul revision
Made several changes to the fundamental way WildeStone works. I removed
the Game class as it was just functioning as a wrapper for Engine and
Board classes.

I changed ascii_lowercase to ascii_uppercase due to board coords needing
to be uppercase on print. I also changed EMPTY to "." and made an
OTHER_SIDE global that points to the other player's color.

In the Board class, I created a history attribute, undo_last_move,
updated move_is_legal to not take color as argument. The main function
that is added is get_neighbors. Also added get_stone and remove_stone.

In the Engine class, I simplified the methods to select_random_move and
make_engine_move, which calls select_random_move and handles if the move
selected was legal or not.

The group class finally has a fleshed out find_all_stones method,
but I may fiddle around with the name of it like "populate_group" or
something.

I also updated both WildeStone.py and test.py to meet all PEP8
guidelines.